feat: add imprint functionality
This commit adds an `imprint` view and adds a route that uses the
`utils.get_imprint` method to add an imprint text to the context of the
view.
It also adds the `apis_acdhch_default_settings` app to the list of
installed apps, so the templates can be picked up automatically.