feat(views): replace utils with views
The `utils` module is a remnant from other Django apps where it
contained additional logic. It now only contains one method for fetching
the imprint. The imprint logic is not used anywhere else, so it can be
part of its own individual view, which is a more Django standardized way
of implementing the imprint functionality.

Closes: #134