build: move, convert project/package settings
Move main project metadata from `[tool.poetry]` section to new
`[project]` section and update the format of individual
settings which require it (like `authors`, `license`).
This change attempts to rectify the breaking of the config
file's formatting when `[project.urls]` was introduced
without an overall `[project]` section.
Not yet included is a conversion of the project dependencies
to the Python/PyPA default format; for the moment, they are
included dynamically.

Resolves: #175